add CmdHelper
