+++ Form1.cs
@@ -62,7 +62,7 @@ namespace ZhengHuo
         {
             //1、断线
             string ccc3 = "netsh advfirewall firewall set rule name=\"ZhengHuo\" new program=\"" + GP + "\" enable=yes >nul";
-            RunCmd(ccc3);
+            CmdHelper.RunCmd(ccc3);
             //2、等N秒
             int seconds = Convert.ToInt32(textBox1.Text)+1;
             pi = seconds;
@@ -80,33 +80,6 @@ namespace ZhengHuo
             Reconnect();
         }
 
-        private static string CmdPath = @"C:\Windows\System32\cmd.exe";
-        public static string RunCmd(string cmd)
-        {
-            cmd = cmd.Trim().TrimEnd('&') + "&exit";//说明：不管命令是否成功均执行exit命令，否则当调用ReadToEnd()方法时，会处于假死状态
-            using (Process p = new Process())
-            {
-                p.StartInfo.FileName = CmdPath;
-                p.StartInfo.UseShellExecute = false;        //是否使用操作系统shell启动
-                p.StartInfo.RedirectStandardInput = true;   //接受来自调用程序的输入信息
-                p.StartInfo.RedirectStandardOutput = true;  //由调用程序获取输出信息
-                p.StartInfo.RedirectStandardError = true;   //重定向标准错误输出
-                p.StartInfo.CreateNoWindow = true;          //不显示程序窗口
-                p.Start();//启动程序
-
-                //向cmd窗口写入命令
-                p.StandardInput.WriteLine(cmd);
-                p.StandardInput.AutoFlush = true;
-
-                //获取cmd窗口的输出信息
-                string output = p.StandardOutput.ReadToEnd();
-                p.WaitForExit();//等待程序执行完退出进程
-                p.Close();
-
-                return output;
-            }
-        }
-
         public static void Reconnect()
         {
             if (pi != 0)
@@ -117,14 +90,14 @@ namespace ZhengHuo
             {
                 timer.Stop();
                 string ccc4 = "netsh advfirewall firewall set rule name=\"ZhengHuo\" new enable=no >nul";
-                RunCmd(ccc4);
+                CmdHelper.RunCmd(ccc4);
                 MessageBox.Show("网络已恢复", "整活");
             }
         }
         private void button1_Click(object sender, EventArgs e)
         {
             string ccc4 = "netsh advfirewall firewall set rule name=\"ZhengHuo\" new enable=no >nul";
-            RunCmd(ccc4);
+            CmdHelper.RunCmd(ccc4);
             MessageBox.Show("网络已恢复", "整活");
         }
     }
